test: fix

+++ Message.cpp
@@ -97,7 +97,7 @@ static void add_amqp_message_annotation(MESSAGE_HANDLE message, AMQP_VALUE msg_a
 }
 
 static std::string* explode(const std::string& str, char delimiter, uint64_t& size) {
-    int count = 1;
+    uint64_t count = 1;
     for (char c : str) {
         if (c == delimiter) {
             count++;
